requirements.txt
# Core ML/DL frameworks (CPU-only)
torch==2.0.1
torchvision==0.15.2
timm==0.9.12

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Visualization
matplotlib==3.8.2

# Streamlit UI
streamlit==1.29.0

# Utility
tqdm==4.66.1